# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FaultTrace-RAG API",
    "description": "Counterfactual Fault Localization for Corpus-Level LLM Analytics Pipelines. Provides deterministic corpus data, procedural queries, dual gold evaluation, and traced pipeline execution.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/runs": {
      "post": {
        "tags": [
          "Runs"
        ],
        "summary": "Execute a pipeline run",
        "description": "Execute a pipeline against a query and store the run.",
        "operationId": "create_run_api_v1_runs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "List pipeline runs",
        "operationId": "list_runs_api_v1_runs_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "query_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query Id"
            }
          },
          {
            "name": "pipeline_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pipeline Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_RunResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/map-plan": {
      "post": {
        "tags": [
          "Runs"
        ],
        "summary": "Dry-run map plan",
        "description": "Generate an extraction plan without executing the pipeline.",
        "operationId": "create_map_plan_api_v1_runs_map_plan_post",
        "parameters": [
          {
            "name": "batch_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Batch Size"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Get run details",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/trace": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Get trace events for a run",
        "operationId": "get_run_trace_api_v1_runs__run_id__trace_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TraceEventResponse"
                  },
                  "title": "Response Get Run Trace Api V1 Runs  Run Id  Trace Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/attribution": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Counterfactual fault attribution for a run",
        "description": "Run counterfactual oracle-replacement attribution on a completed run.\n\nReturns Shapley values and REF scores for each pipeline component:\n- scope (R): which records were retrieved\n- facts (E): which field values were extracted\n- aggregation (A): how facts were reduced to an answer",
        "operationId": "get_run_attribution_api_v1_runs__run_id__attribution_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/batch-attribution": {
      "post": {
        "tags": [
          "Runs"
        ],
        "summary": "Batch counterfactual fault attribution",
        "description": "Run counterfactual oracle-replacement attribution on a batch of completed runs.",
        "operationId": "batch_attribution_api_v1_runs_batch_attribution_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "title": "Run Ids"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/runs/{run_id}/certificate": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Get certification result for a run",
        "operationId": "get_run_certificate_api_v1_runs__run_id__certificate_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/batch-evaluate-policy": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Evaluate a policy across all completed runs",
        "operationId": "batch_evaluate_policy_api_v1_runs_batch_evaluate_policy_get",
        "parameters": [
          {
            "name": "policy_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "strict_exact_v1",
              "title": "Policy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRunRequest": {
        "properties": {
          "query_id": {
            "type": "string",
            "title": "Query Id"
          },
          "pipeline_id": {
            "type": "string",
            "title": "Pipeline Id",
            "default": "P0-deterministic-scope-baseline"
          },
          "provider_id": {
            "type": "string",
            "title": "Provider Id",
            "default": "deterministic"
          }
        },
        "type": "object",
        "required": [
          "query_id"
        ],
        "title": "CreateRunRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_RunResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "has_next"
        ],
        "title": "PaginatedResponse[RunResponse]"
      },
...
      "RunResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "query_id": {
            "type": "string",
            "title": "Query Id"
          },
          "pipeline_id": {
            "type": "string",
            "title": "Pipeline Id"
          },
          "provider_id": {
            "type": "string",
            "title": "Provider Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "gold_answer_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gold Answer Value"
          },
          "is_correct": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Correct"
          },
          "loss": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loss"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "config_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Hash"
          },
          "artifact_refs": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Artifact Refs",
            "default": {}
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "query_id",
          "pipeline_id",
          "provider_id",
          "status",
          "started_at"
        ],
        "title": "RunResponse"
      },
...
      "TraceEventResponse": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "parent_event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Event Id"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "record_count_in": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Count In"
          },
          "record_count_out": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Count Out"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload",
            "default": {}
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "run_id",
          "stage",
          "event_type",
          "message",
          "timestamp"
        ],
        "title": "TraceEventResponse"
      },
...
    }
  }
}